requirements.txt
langchain
openai
huggingface_hub
tensorflow
langchain-openai
python-dotenv